升级mybatis-plus为3.5.1版本 删除冲突的mybatis包 新增角色管理页面相关接口与页面

+++ src/main/java/com/book/common/units/PageInfo.java
@@ -12,7 +12,7 @@ import java.util.Map;
 public class PageInfo {
 
     private final static int PAGESIZE = 10; //默认显示的记录数 
-    private int total; // 总记录 
+    private long total; // 总记录
     private List rows; //显示的记录  
     @JsonIgnore
     private int from;
@@ -77,11 +77,11 @@ public class PageInfo {
         this.order = order;
     }
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(long total) {
         this.total = total;
     }
